+++ sub-packages/bionemo-scspeedtest/pyproject.toml
@@ -3,8 +3,8 @@ requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "bionemo-scbenchmark"
-version = "0.1.0"
+name = "bionemo-scspeedtest"
+version = "0.0.1"
 description = "Simple benchmarking framework for any single-cell dataloader"
 readme = "README.md"
 requires-python = ">=3.8"
